Added wishlist component
